+++ README.md
@@ -33,3 +33,12 @@
 ## 예외 처리
 - [X] 요청 메세지 값의 null, 공백 검증
 
+
+## Todo
+- 테스트 코드 작성
+  - 통합 테스트 작성
+    - 컨트롤러 단위로 나눠서 통합 테스트 작성
+- [X] 리팩토링
+  - [X] 컨트롤러 분리
+  - [X] 예외처리 객체 분리
+
