+++ src/main/java/cart/controller/AdminController.java
@@ -1,0 +1,56 @@
+package cart.controller;
+
+import cart.dto.request.ProductRequest;
+import cart.dto.request.ProductUpdateRequest;
+import cart.dto.response.ProductResponse;
+import cart.service.CartService;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+
+import javax.validation.Valid;
+import java.util.List;
+
+@Controller
+@RequestMapping("/admin")
+public class AdminController {
+
+    private final CartService cartService;
+
+    public AdminController(CartService cartService) {
+        this.cartService = cartService;
+    }
+
+    @GetMapping("")
+    public String admin(Model model) {
+        List<ProductResponse> productsResponse = cartService.readAll();
+        model.addAttribute("products", productsResponse);
+        return "admin";
+    }
+
+    @PostMapping("/create")
+    public ResponseEntity<String> create(@RequestBody @Valid ProductRequest productRequest) {
+        cartService.create(productRequest);
+        return ResponseEntity.ok().build();
+    }
+
+    @PutMapping("/update")
+    public ResponseEntity<String> update(@RequestBody @Valid ProductUpdateRequest productUpdateRequest) {
+        cartService.update(productUpdateRequest);
+        return ResponseEntity.ok().build();
+    }
+
+    @DeleteMapping("/delete")
+    public ResponseEntity<String> delete(@RequestParam Long id) {
+        System.out.println(id);
+        cartService.delete(id);
+        return ResponseEntity.ok().build();
+    }
+}
