+++ src/main/java/cart/controller/CartController.java
@@ -1,26 +1,12 @@
 package cart.controller;
 
-import cart.dto.request.ProductRequest;
-import cart.dto.request.ProductUpdateRequest;
 import cart.dto.response.ProductResponse;
 import cart.service.CartService;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
 
-import javax.validation.Valid;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 @Controller
 public class CartController {
@@ -36,43 +22,5 @@ public class CartController {
         List<ProductResponse> productsResponse = cartService.readAll();
         model.addAttribute("products", productsResponse);
         return "index";
-    }
-
-    @GetMapping("/admin")
-    public String admin(Model model) {
-        List<ProductResponse> productsResponse = cartService.readAll();
-        model.addAttribute("products", productsResponse);
-        return "admin";
-    }
-
-    @PostMapping("/admin/create")
-    public ResponseEntity<String> create(@RequestBody @Valid ProductRequest productRequest) {
-        cartService.create(productRequest);
-        return ResponseEntity.ok().build();
-    }
-
-    @PutMapping("/admin/update")
-    public ResponseEntity<String> update(@RequestBody @Valid ProductUpdateRequest productUpdateRequest) {
-        cartService.update(productUpdateRequest);
-        return ResponseEntity.ok().build();
-    }
-
-    @DeleteMapping("/admin/delete")
-    public ResponseEntity<String> delete(@RequestParam Long id) {
-        System.out.println(id);
-        cartService.delete(id);
-        return ResponseEntity.ok().build();
-    }
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<Map<String, String>> methodArgumentNotValidExceptionHandler(
-            MethodArgumentNotValidException exception) {
-        Map<String, String> errors = new HashMap<>();
-        exception.getBindingResult().getAllErrors().forEach((error) -> {
-            String fieldName = ((FieldError) error).getField();
-            String errorMessage = error.getDefaultMessage();
-            errors.put(fieldName, errorMessage);
-        });
-        return ResponseEntity.badRequest().body(errors);
     }
 }
